Extract PersonalFinanceCategoryEnum from generated Plaid client
The PersonalFinanceCategoryEnum existed as custom code that had
been embedded into the generated Plaid client. This extracts the
custom enum and associated logic so that the Plaid client may be
more easily regenerated from the OpenAPI spec.
